Refactors: - Int to Double en stats y atributos. - Ya no se genera la tabla aventurero_tacticas, la tactica conoce al aventurero que la generó, si se elimina el aventurero persistido, la tactica persistida también. - Ahora CURAR quita el dañoRecibido y no aumenta la vida, así puede reflejarse en el front y el aventurero no tiene vida de más.

+++ src/main/kotlin/ar/edu/unq/epers/tactics/modelo/enums/Criterio.kt
@@ -1,9 +1,9 @@
 package ar.edu.unq.epers.tactics.modelo.enums
 
 enum class Criterio {
-    IGUAL { override fun evaluarseCon(valorAComparar: Int, valorDeComparacion: Int) = valorAComparar == valorDeComparacion },
-    MAYOR_QUE { override fun evaluarseCon(valorAComparar: Int, valorDeComparacion: Int) = valorAComparar > valorDeComparacion },
-    MENOR_QUE { override fun evaluarseCon(valorAComparar: Int, valorDeComparacion: Int) = valorAComparar < valorDeComparacion };
+    IGUAL { override fun evaluarseCon(valorAComparar: Double, valorDeComparacion: Double) = valorAComparar == valorDeComparacion },
+    MAYOR_QUE { override fun evaluarseCon(valorAComparar: Double, valorDeComparacion: Double) = valorAComparar > valorDeComparacion },
+    MENOR_QUE { override fun evaluarseCon(valorAComparar: Double, valorDeComparacion: Double) = valorAComparar < valorDeComparacion };
 
-    abstract fun evaluarseCon(valorAComparar: Int, valorDeComparacion: Int): Boolean
+    abstract fun evaluarseCon(valorAComparar: Double, valorDeComparacion: Double): Boolean
 }
